% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_operations.R
\name{add_to_logFile}
\alias{add_to_logFile}
\title{Add DB modifications to a log entry}
\usage{
add_to_logFile(what, who, info)
}
\arguments{
\item{what}{Action taken}

\item{who}{Who did it. The user}

\item{info}{A list of the information to print on the log file}
}
\value{
Saves on the DB
}
\description{
Add DB modifications to a log entry
}
