% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_information.R
\name{mod_sample_information_server}
\alias{mod_sample_information_server}
\title{sample_information Server Functions}
\usage{
mod_sample_information_server(id)
}
\description{
sample_information Server Functions
}
