% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_storage_information.R
\name{mod_storage_information_server}
\alias{mod_storage_information_server}
\title{storage_information Server Functions}
\usage{
mod_storage_information_server(id, sample_info)
}
\description{
storage_information Server Functions
}
