% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_placeUI.R
\name{storage_placeUI}
\alias{storage_placeUI}
\title{UI for the Storage Place}
\usage{
storage_placeUI(ns)
}
\description{
This is a function that returns the UI for the Storage Place of a specimen
It is used in both the adding a specimen for the first time, and also for modifying the
specimen
}
