% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_information_fct_utils.R
\name{to_date_time}
\alias{to_date_time}
\title{Get date out of numeric format time}
\usage{
to_date_time(number, .tz = "EET")
}
\arguments{
\item{number}{A number that needs be translated to date}

\item{.tz}{Time zone}
}
\description{
Translates a number to a date. 
Considers origin the 01/01/1970
}
