% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_operations.R
\name{update_duration}
\alias{update_duration}
\title{Update the duration from sample collection to sample processing}
\usage{
update_duration(conn, specimen, user)
}
\arguments{
\item{conn}{the connection to the DB}

\item{specimen}{the specimen list}
}
\description{
Calculates and updates to the database
}
\details{
This function will usually be called when the user changes either the 
date of collection or the date of processing
}
